Fixed test and rubocop
