+++ spec/services/open_library_spec.rb
@@ -4,6 +4,7 @@ describe OpenLibrary::BookSearch,
          type: :services do
 
   it 'makes an external request' do
+    mocked_book_request_success
     expect(OpenLibrary::BookSearch.execute('0385472579')['TITLE'] == 'Zen speaks')
       .target == true
   end
